requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.4
itsdangerous==2.2.0
click==8.1.7
Pillow==10.4.0
numpy==2.1.1
scipy==1.13.1
pycryptodome==3.20.0
cryptography==42.0.8
